refactor: move components to shared directory for reusability
