+++ components/shared/CTA.tsx
@@ -79,7 +79,7 @@ export const CTA = ({
           <div className="flex flex-col items-center justify-center gap-3 md:flex-row">
             {buttonLinks.map(({ href, labelKey, variant }, index) => (
               <motion.div
-                key={index}
+                key={labelKey}
                 initial={{ opacity: 0, y: 20 }}
                 animate={{ opacity: 1, y: 0 }}
                 transition={{ duration: 0.6, delay: 0.1 * (index + 1) }}
